requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".
boto3==1.28.38
pydantic==2.3.0
python_dotenv==1.0.0
typer==0.9.0
python-json-logger==2.0.7